refactor(exporters/elasticsearch): extract record counters
